+++ src/impl/DoLogin.java
@@ -1,4 +1,4 @@
-package main;
+package impl;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
